refactor(#412): extract renderContent to components/renderContent.tsx (-48 lines)
