+++ apps/desktop/src/renderer/features/chat/ChatConsole.tsx
@@ -3454,51 +3454,4 @@ function MessageBubble({
  *  Handles both complete blocks and cross-message orphans
  *  (tags split across streaming chunks). */
 
-function renderContent(text: string) {
-  const parts = text.split(/(```[\s\S]*?```)/g);
-  return parts.map((part, i) => {
-    if (part.startsWith('```') && part.endsWith('```')) {
-      const inner = part.slice(3, -3);
-      const langEnd = inner.indexOf('\n');
-      const code = langEnd > 0 ? inner.slice(langEnd + 1) : inner;
-      return (
-        <pre
-          key={i}
-          className="my-2 text-xs rounded-lg px-3 py-2 overflow-x-auto"
-          style={{ background: 'rgba(0,0,0,0.06)' }}
-        >
-          <code>{code}</code>
-        </pre>
-      );
-    }
-    const segments = part.split(/(\*\*[^*]+\*\*|`[^`]+`)/g);
-    return (
-      <span key={i}>
-        {segments.map((seg, j) => {
-          if (seg.startsWith('**') && seg.endsWith('**'))
-            return <strong key={j}>{seg.slice(2, -2)}</strong>;
-          if (seg.startsWith('`') && seg.endsWith('`'))
-            return (
-              <code
-                key={j}
-                className="text-xs font-mono px-1 rounded"
-                style={{ background: 'rgba(0,0,0,0.08)' }}
-              >
-                {seg.slice(1, -1)}
-              </code>
-            );
-          return (
-            <span key={j}>
-              {seg.split('\n').map((line, k, arr) => (
-                <span key={k}>
-                  {line}
-                  {k < arr.length - 1 && <br />}
-                </span>
-              ))}
-            </span>
-          );
-        })}
-      </span>
-    );
-  });
-}
+import { renderContent } from './components/renderContent';
